refactor: 🚚 organized Hero Content
Follows `/components` folder structure

Related: #179
