+++ src/components/Main/Hero/index.tsx
